When a usage and project both exist, change how dependencies are consumed
If a project and usage exist, use the project first, then the usage.

If a project uses itself, any link dependencies in the usage will be consumed by the project.
Otherwise those link dependencies will pass through to the user.

This feels a lot more sensible, and should remove the need for link/build option filtering.